Allow disabling the server hostname check when using SSL
When running tests over HTTPS on changing machines, having a certificate
matching the hostname of each machine quickly becomes cumbersome. It
was already possible to accept a certificate signed by any authority by
adding it to the keystore, but RestAssured would not let us disable the
hostname check.

This commit adds a new form of RestAssured.certificate() with a boolean
parameter allowing to turn the hostname check on or off. By default, the
check is still on, to preserve backwards compatibility.